Moving address, abi and type to its own files
