+++ examples/4337-gas-metering/alchemy/alchemy.ts
@@ -5,7 +5,8 @@ import { setTimeout } from 'timers/promises'
 import { PublicClient, Hash, Transport, createPublicClient, formatEther, http, parseEther, zeroAddress } from 'viem'
 import { privateKeyToAccount } from 'viem/accounts'
 import { goerli, sepolia } from 'viem/chains'
-import { SAFE_ADDRESSES_MAP, getAccountAddress, getAccountInitCode } from '../utils/safe'
+import { getAccountAddress, getAccountInitCode } from '../utils/safe'
+import { SAFE_ADDRESSES_MAP } from '../utils/address'
 import {
   UserOperation,
   signUserOperation,
