+++ examples/4337-gas-metering/gelato/gelato.ts
@@ -2,7 +2,8 @@ import dotenv from 'dotenv'
 import { Hash, createPublicClient, http, zeroAddress } from 'viem'
 import { privateKeyToAccount } from 'viem/accounts'
 import { baseSepolia, sepolia } from 'viem/chains'
-import { SAFE_ADDRESSES_MAP, getAccountAddress, getGelatoAccountInitCode, getGelatoCallData, prepareForGelatoTx } from '../utils/safe'
+import { getAccountAddress, getGelatoAccountInitCode, getGelatoCallData, prepareForGelatoTx } from '../utils/safe'
+import { SAFE_ADDRESSES_MAP } from '../utils/address'
 import { txTypes } from '../utils/userOps'
 import { GelatoRelay } from '@gelatonetwork/relay-sdk'
 import { setTimeout } from 'timers/promises'
