+++ examples/4337-gas-metering/pimlico/pimlico.ts
@@ -5,7 +5,8 @@ import { setTimeout } from 'timers/promises'
 import { Client, Hash, createClient, createPublicClient, http } from 'viem'
 import { privateKeyToAccount } from 'viem/accounts'
 import { goerli, polygonMumbai } from 'viem/chains'
-import { SAFE_ADDRESSES_MAP, getAccountAddress, getAccountInitCode } from '../utils/safe'
+import { getAccountAddress, getAccountInitCode } from '../utils/safe'
+import { SAFE_ADDRESSES_MAP } from '../utils/address'
 import { UserOperation, submitUserOperationPimlico, signUserOperation, txTypes, createCallData } from '../utils/userOps'
 import { getERC20Decimals, getERC20Balance, transferERC20Token } from '../utils/erc20'
