+++ examples/4337-gas-metering/utils/type.ts
@@ -1,0 +1,34 @@
+export const EIP712_SAFE_TX_TYPE = {
+    // "SafeTx(address to,uint256 value,bytes data,uint8 operation,uint256 safeTxGas,uint256 baseGas,uint256 gasPrice,address gasToken,address refundReceiver,uint256 nonce)"
+    SafeTx: [
+      { type: 'address', name: 'to' },
+      { type: 'uint256', name: 'value' },
+      { type: 'bytes', name: 'data' },
+      { type: 'uint8', name: 'operation' },
+      { type: 'uint256', name: 'safeTxGas' },
+      { type: 'uint256', name: 'baseGas' },
+      { type: 'uint256', name: 'gasPrice' },
+      { type: 'address', name: 'gasToken' },
+      { type: 'address', name: 'refundReceiver' },
+      { type: 'uint256', name: 'nonce' },
+    ],
+  }
+
+
+export const EIP712_SAFE_OPERATION_TYPE = {
+  SafeOp: [
+    { type: 'address', name: 'safe' },
+    { type: 'uint256', name: 'nonce' },
+    { type: 'bytes', name: 'initCode' },
+    { type: 'bytes', name: 'callData' },
+    { type: 'uint256', name: 'callGasLimit' },
+    { type: 'uint256', name: 'verificationGasLimit' },
+    { type: 'uint256', name: 'preVerificationGas' },
+    { type: 'uint256', name: 'maxFeePerGas' },
+    { type: 'uint256', name: 'maxPriorityFeePerGas' },
+    { type: 'bytes', name: 'paymasterAndData' },
+    { type: 'uint48', name: 'validAfter' },
+    { type: 'uint48', name: 'validUntil' },
+    { type: 'address', name: 'entryPoint' },
+  ],
+}
