+++ examples/4337-gas-metering/utils/userOps.ts
@@ -1,7 +1,8 @@
 import dotenv from 'dotenv'
 import type { Address } from 'abitype'
 import { fromHex, parseEther, type Hex, type PrivateKeyAccount, formatEther } from 'viem'
-import { EIP712_SAFE_OPERATION_TYPE, encodeCallData } from './safe'
+import { encodeCallData } from './safe'
+import { EIP712_SAFE_OPERATION_TYPE } from './type'
 import { Alchemy } from 'alchemy-sdk'
 import { setTimeout } from 'timers/promises'
 import { generateTransferCallData, getERC20Balance, getERC20Decimals, mintERC20Token } from './erc20'
